toolkit/components/extensions/test/mochitest/test_ext_schema.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for schema API creation</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SpawnTask.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/ExtensionTestUtils.js"></script>
  <script type="text/javascript" src="head.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>

<script type="text/javascript">
"use strict";

add_task(function* testEmptySchema() {
  function background() {
    browser.test.assertTrue(!("manifest" in browser), "browser.manifest is not defined");
    browser.test.assertTrue("storage" in browser, "browser.storage should be defined");
    browser.test.assertTrue(!("contextMenus" in browser), "browser.contextMenus should not be defined");
    browser.test.notifyPass("schema");
  }

  let extension = ExtensionTestUtils.loadExtension({
    background: `(${background})()`,
    manifest: {
      permissions: ["storage"],
    },
  });

  yield extension.startup();
  yield extension.awaitFinish("schema");
  yield extension.unload();
});

add_task(function* testUnknownProperties() {
  function background() {
    browser.test.notifyPass("loaded");
  }

  let extension = ExtensionTestUtils.loadExtension({
    manifest: {
      permissions: ["unknownPermission"],

      unknown_property: {},
    },

    background: `(${background})()`,
  });

  let messages = [
    {message: /processing permissions\.0: Unknown permission "unknownPermission"/},
    {message: /processing unknown_property: An unexpected property was found in the WebExtension manifest/},
  ];

  let waitForConsole = new Promise(resolve => {
    SimpleTest.monitorConsole(resolve, messages);
  });

  yield extension.startup();

  yield extension.awaitFinish("loaded");

  yield extension.unload();

  SimpleTest.endMonitorConsole();
  yield waitForConsole;
});
</script>

</body>
</html>
